refactor lamp exceptions and move to runtime_utils. add new exceptions for expectedinputsnotfound and invalidprocessing
