+++ src/lamp_py/ingestion/convert_gtfs_rt.py
@@ -43,7 +43,7 @@ from lamp_py.ingestion.config_busloc_vehicle import RtBusVehicleDetail
 from lamp_py.ingestion.config_rt_trip import RtTripDetail
 from lamp_py.ingestion.config_rt_vehicle import RtVehicleDetail
 from lamp_py.ingestion.converter import ConfigType, Converter
-from lamp_py.ingestion.error import NoImplException
+from lamp_py.runtime_utils.lamp_exception import NoImplException
 from lamp_py.ingestion.gtfs_rt_detail import GTFSRTDetail
 from lamp_py.ingestion.utils import (
     GTFS_RT_HASH_COL,
